separated world collisions from square collisions
